completely rewrote the `DownloadManager` ( now is `UploadService` )
main changes:
------------
 - now the Gentoo root is in the application data
 - Downloads are now performed in background
 - we can now extract tarballs
 - we can now write to spawned processes stdin
 - added some stuff about Gentoo root in preferences

enjoy :smirk:

closes #307, #308, #317, #319

+++ dSploit/src/it/evilsocket/dsploit/core/Shell.java
@@ -18,14 +18,13 @@
  */
 package it.evilsocket.dsploit.core;
 
-import android.content.Context;
-
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.UUID;
 
@@ -38,19 +37,23 @@ public class Shell
   private final static ArrayList<Integer> mFIFOS = new ArrayList<Integer>();
   private static Thread fakeThread;
 
-  /*
+  /**
    * "gobblers" seem to be the recommended way to ensure the streams
-   * don't cause issues
+   * don't cause issues.
+   * this class keeps reading from process output until they close it.
    */
   public static class StreamGobbler extends Thread {
     private BufferedReader mReader = null;
     private OutputReceiver mReceiver = null;
+    private OutputStream mOutputStream = null;
     private String mToken = null;
     private String mFIFOPath = null;
+    private String mOutputFIFOPath = null;
     private int mFIFONum = -1;
+    private int mOutputFIFONum = -1;
     public int exitValue = -1;
 
-    public StreamGobbler(String fifoPath, OutputReceiver receiver, String token, int fifoNum) throws IOException {
+    public StreamGobbler(String fifoPath, OutputReceiver receiver, String token, int fifoNum) {
       mFIFOPath = fifoPath;
       mReceiver = receiver;
       mToken = token;
@@ -58,16 +61,30 @@ public class Shell
       setDaemon(true);
     }
 
+    public void setOutputFifo(String fifoPath, int fifoNum) {
+      mOutputFIFONum = fifoNum;
+      mOutputFIFOPath = fifoPath;
+    }
+
+    public OutputStream getOutputStream(){
+      return mOutputStream;
+    }
+
     public void run(){
       try
       {
-        String line = null;
+        String line;
         // wait for the FIFO to be created
         while(!(new File(mFIFOPath).exists()))
           Thread.sleep(200);
+        if(mOutputFIFOPath!=null)
+          while(!(new File(mOutputFIFOPath)).exists())
+            Thread.sleep(200);
 
         // this will hang until someone will connect to the other side of the pipe.
         mReader = new BufferedReader(new FileReader(mFIFOPath));
+        if(mOutputFIFOPath!=null)
+          mOutputStream = new FileOutputStream(mOutputFIFOPath);
 
         while((line=mReader.readLine())!=null) {
           if(line.startsWith(mToken)) {
@@ -89,18 +106,22 @@ public class Shell
         try{
           if(mReader!=null)
             mReader.close();
+          if(mOutputStream!=null)
+            mOutputStream.close();
           synchronized (mFIFOS) {
             mFIFOS.remove(mFIFOS.indexOf(mFIFONum));
+            if(mOutputFIFONum>=0)
+              mFIFOS.remove(mFIFOS.indexOf(mOutputFIFONum));
           }
         } catch(IOException e){
-          //swallow error
+          System.errorLogging(e);
         }
       }
     }
   }
 
-  private static Process spawnShell(String command) throws IOException {
-    return new ProcessBuilder().command(command).start();
+  private static Process spawnSuShell() throws IOException {
+    return new ProcessBuilder().command("su").start();
   }
 
   public static boolean isRootGranted() {
@@ -131,6 +152,7 @@ public class Shell
 
   private static void clearFifos() throws IOException {
     File d = new File(System.getFifosPath());
+    File[] files;
     if(!d.exists())
       d.mkdir();
 
@@ -138,8 +160,8 @@ public class Shell
       d.delete();
       d.mkdir();
     }
-    if(d.isDirectory()) {
-      for(File f : d.listFiles()) {
+    if(d.isDirectory() && (files = d.listFiles())!=null) {
+      for(File f : files) {
         f.delete();
       }
     }
@@ -148,22 +170,51 @@ public class Shell
     }
   }
 
-  public static int exec(String command, OutputReceiver receiver) throws IOException, InterruptedException {
-    Thread g = async(command, receiver);
+  /**
+   * execute a command in a root shell
+   * @param command    the command to execute
+   * @param receiver   the {@link it.evilsocket.dsploit.core.Shell.OutputReceiver} that will receive command output
+   * @param withInput  true if should create an output stream connected to the command stdin
+   * @return the exit code of the command
+   * @throws IOException if cannot create files, streams or processes
+   * @throws InterruptedException when interrupt is received
+   */
+  public static int exec(String command, OutputReceiver receiver, boolean withInput) throws IOException, InterruptedException {
+    Thread g = async(command, receiver, withInput);
     if(g==fakeThread)
-      return -1;
+      throw new IOException("cannot execute shell commands");
     g.start();
     g.join();
     return  ((StreamGobbler)g).exitValue;
   }
 
-  public static int exec(String command) throws IOException, InterruptedException {
-    return exec(command, null);
+  /**
+   * {@code withInput} default to false
+   * @see
+   */
+  public static int exec(String command, OutputReceiver receiver) throws IOException, InterruptedException {
+    return exec(command, receiver, false);
   }
 
-  public static Thread async(final String command, final OutputReceiver receiver) {
-    int fifonum;
-    String fifo_path;
+  /**
+   * {@inheritDoc exec}
+   */
+  public static int exec(String command) throws IOException, InterruptedException {
+    return exec(command, null, false);
+  }
+
+  /**
+   * start command in a root shell now and read it's output asynchronously
+   * @param command    the command to execute
+   * @param receiver   the {@link it.evilsocket.dsploit.core.Shell.OutputReceiver} that will receive command output
+   * @param withInput  true if should create an output stream connected to the command stdin
+   * @return {@link it.evilsocket.dsploit.core.Shell.StreamGobbler} that keep reading from shell output until process dies,
+   *          a fakeThread on error.
+   */
+  @SuppressWarnings("StatementWithEmptyBody")
+  public static Thread async(final String command, final OutputReceiver receiver, boolean withInput) {
+    int fifonum,inputFifoNum;
+    String fifo_path,input_fifo_path;
     String token = UUID.randomUUID().toString().toUpperCase();
     StreamGobbler gobbler = null;
 
@@ -171,7 +222,7 @@ public class Shell
     {
       // spawn shell for the first time
       if(mRootShell==null) {
-        mRootShell = spawnShell("su");
+        mRootShell = spawnSuShell();
         mWriter = new DataOutputStream(mRootShell.getOutputStream());
         // this 2 reader are useful for debugging purposes
         //reader = new BufferedReader(new InputStreamReader(mRootShell.getInputStream()));
@@ -186,20 +237,39 @@ public class Shell
       {
         // check for next available FIFO number
         for( fifonum = 0; fifonum < MAX_FIFO && mFIFOS.contains(fifonum); fifonum++ );
-        if(fifonum==1024)
+        if(fifonum==MAX_FIFO)
           throw new IOException("maximum number of fifo reached");
 
         mFIFOS.add(fifonum);
+
+        if(withInput) {
+          for(inputFifoNum=0; inputFifoNum < MAX_FIFO && mFIFOS.contains(inputFifoNum); inputFifoNum++);
+          if(inputFifoNum==MAX_FIFO)
+            throw new IOException("maximum number of fifo reached");
+          mFIFOS.add(inputFifoNum);
+        } else
+          inputFifoNum=-1;
       }
 
       fifo_path = String.format("%s%d",System.getFifosPath(),fifonum);
 
       mWriter.writeBytes(String.format("mkfifo -m 666 '%s'\n", fifo_path));
-      mWriter.flush();
-      mWriter.writeBytes(String.format("(%s%s echo -e \"\\n%s$?\") 2>&1 >%s &\n", command,(command.trim().endsWith(";") ? "" : " ;"), token, fifo_path));
+      if(inputFifoNum>=0) {
+        input_fifo_path = String.format("%s%d",System.getFifosPath(), inputFifoNum);
+        mWriter.writeBytes(String.format("mkfifo -m 666 '%s'\n", input_fifo_path));
+      } else
+        input_fifo_path = null;
+      mWriter.writeBytes(String.format("(%s%s echo -e \"\\n%s$?\") 2>&1 >%s %s &\n",
+                          command,
+                          (command.trim().endsWith(";") ? "" : " ;"),
+                          token,
+                          fifo_path,
+                          (inputFifoNum>=0 ? "<"+input_fifo_path : "")));
       mWriter.flush();
 
       gobbler = new StreamGobbler(fifo_path, receiver, token, fifonum);
+      if(inputFifoNum>=0)
+        gobbler.setOutputFifo(input_fifo_path, inputFifoNum);
     }
     catch ( IOException e) {
       System.errorLogging(e);
@@ -218,6 +288,19 @@ public class Shell
       return gobbler;
   }
 
+  /**
+   * calls {@link #async(String, it.evilsocket.dsploit.core.Shell.OutputReceiver, boolean)}
+   * with {@code withInput} to false
+   */
+  public static Thread async(String command, OutputReceiver receiver) {
+    return async(command, receiver, false);
+  }
+
+  /**
+   * calls {@link #async(String, it.evilsocket.dsploit.core.Shell.OutputReceiver, boolean)}
+   * with null {@link it.evilsocket.dsploit.core.Shell.OutputReceiver}
+   * and {@code withInput} to false
+   */
   public static Thread async(String command) {
     return async(command, null);
   }
